+++ src/client.ts
@@ -14,6 +14,7 @@ import { WSConnectionFallback } from './connection_fallback';
 import { isErrorResponse, isWSFailure } from './errors';
 import {
   addFileToFormData,
+  axiosParamsSerializer,
   chatCodes,
   isFunction,
   isOnline,
@@ -155,6 +156,7 @@ import {
   UpdateCommandResponse,
   UpdatedMessage,
   UpdateMessageAPIResponse,
+  UpdateMessageOptions,
   UserCustomEvent,
   UserFilters,
   UserOptions,
@@ -303,6 +305,8 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
 
     this.defaultWSTimeoutWithFallback = 6000;
     this.defaultWSTimeout = 15000;
+
+    this.axiosInstance.defaults.paramsSerializer = axiosParamsSerializer;
 
     /**
      * logger function should accept 3 parameters:
@@ -1460,7 +1464,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
    *
    * @param {BannedUsersFilters} filterConditions MongoDB style filter conditions
    * @param {BannedUsersSort} sort Sort options [{created_at: 1}].
-   * @param {BannedUsersPaginationOptions} options Option object, {limit: 10, offset:0}
+   * @param {BannedUsersPaginationOptions} options Option object, {limit: 10, offset:0, exclude_expired_bans: true}
    *
    * @return {Promise<BannedUsersResponse<StreamChatGenerics>>} Ban Query Response
    */
@@ -1590,8 +1594,8 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     query: string | MessageFilters<StreamChatGenerics>,
     options: SearchOptions<StreamChatGenerics> = {},
   ) {
-    if (options.offset && (options.sort || options.next)) {
-      throw Error(`Cannot specify offset with sort or next parameters`);
+    if (options.offset && options.next) {
+      throw Error(`Cannot specify offset with next`);
     }
     const payload: SearchPayload<StreamChatGenerics> = {
       filter_conditions: filterConditions,
@@ -1742,16 +1746,21 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
       throw Error(`Invalid channel group ${channelType}, can't contain the : character`);
     }
 
+    // support channel("messaging", {options})
+    if (channelIDOrCustom && typeof channelIDOrCustom === 'object') {
+      return this.getChannelByMembers(channelType, channelIDOrCustom);
+    }
+
+    // // support channel("messaging", undefined, {options})
+    if (!channelIDOrCustom && typeof custom === 'object' && custom.members?.length) {
+      return this.getChannelByMembers(channelType, custom);
+    }
+
     // support channel("messaging", null, {options})
     // support channel("messaging", undefined, {options})
     // support channel("messaging", "", {options})
-    if (channelIDOrCustom == null || channelIDOrCustom === '') {
+    if (!channelIDOrCustom) {
       return new Channel<StreamChatGenerics>(this, channelType, undefined, custom);
-    }
-
-    // support channel("messaging", {options})
-    if (typeof channelIDOrCustom === 'object') {
-      return this.getChannelByMembers(channelType, channelIDOrCustom);
     }
 
     return this.getChannelById(channelType, channelIDOrCustom, custom);
@@ -2253,23 +2262,21 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
   }
 
   /**
-   * _unblockMessage - unblocks message blocked by automod
+   * unblockMessage - unblocks message blocked by automod
    *
-   * Note: Do not use this.
-   * It is present for internal usage only.
-   * This function can, and will, break and/or be removed at any point in time.
    *
-   * @private
    * @param {string} targetMessageID
    * @param {string} [options.user_id] currentUserID, only used with serverside auth
    * @returns {Promise<APIResponse>}
    */
-  async _unblockMessage(targetMessageID: string, options: { user_id?: string } = {}) {
+  async unblockMessage(targetMessageID: string, options: { user_id?: string } = {}) {
     return await this.post<APIResponse>(this.baseURL + '/moderation/unblock_message', {
       target_message_id: targetMessageID,
       ...options,
     });
   }
+  // alias for backwards compatibility
+  _unblockMessage = this.unblockMessage;
 
   /**
    * @deprecated use markChannelsRead instead
@@ -2446,7 +2453,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
   async updateMessage(
     message: UpdatedMessage<StreamChatGenerics>,
     userId?: string | { id: string },
-    options?: { skip_enrich_url?: boolean },
+    options?: UpdateMessageOptions,
   ) {
     if (!message.id) {
       throw Error('Please specify the message id when calling updateMessage');
@@ -2517,7 +2524,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     id: string,
     partialMessageObject: PartialMessageUpdate<StreamChatGenerics>,
     userId?: string | { id: string },
-    options?: { skip_enrich_url?: boolean },
+    options?: UpdateMessageOptions,
   ) {
     if (!id) {
       throw Error('Please specify the message id when calling partialUpdateMessage');
@@ -2545,7 +2552,9 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
   }
 
   async getMessage(messageID: string) {
-    return await this.get<GetMessageAPIResponse<StreamChatGenerics>>(this.baseURL + `/messages/${messageID}`);
+    return await this.get<GetMessageAPIResponse<StreamChatGenerics>>(
+      this.baseURL + `/messages/${encodeURIComponent(messageID)}`,
+    );
   }
 
   getUserAgent() {